move otaclient._utils.linux into otaclient_common package
