+++ src/otaclient_common/linux.py
